Use static scrape config to federate from prometheus-cache
The scrape config to federate metrics from prometheus-cache to the Shoot Prometheus instances used Kubernetes service discovery to find the Prometheus service and required access to service resources in the garden namespace. This commit uses the DNS record with a static scrape config to not require any RBAC permissions outside of the shoot specific namespace.