Dedicated cert-based kubeconfig for virtual garden
Without this, when requesting admin/viewer kubeconfigs for a Shoot in
this operator-managed garden, the issued client certificate will always
contain the user name `system:serviceaccount:kube-system:gardener`.
This is because this ServiceAccount was used to request the admin/viewer
kubeconfig.
Since exactly this ServiceAccount is also used by `gardenlet` when
communicating with Shoots, it has cluster-admin access in all shoot
clusters.
This results in viewer kubeconfigs also having cluster-admin access,
breaking the e2e tests.

See for example: https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/12213/pull-gardener-e2e-kind-ha-multi-zone/1929570125401296896

With a dedicated cert-based kubeconfig, we also improve the developers'
lives. The token-based kubeconfig is rotated each `12h`, which is
annoying for development scenarios. This cert-based kubeconfig is valid
for `1y`.