Rework `Garden`, extensions, garden config, and `Gardenlet` kustomizations (move to `/dev-setup`) (#12213)
* Nit: Don't log errors that are returned

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* [provider-local] Runtime webhook config must have different name

Otherwise, if another instance `gardener-extension-provider-local` comes
up later when a `Seed` is registered (i.e., when the garden runtime
cluster is a seed at the same time), they will both fight over the same
`MutatingWebhookConfiguration/gardener-extension-provider-local`.
Concretely, they both try to write their CA bundle into it. The CA
bundles are different, hence, this can lead to errors like this:

```
  - lastTransitionTime: "2025-05-26T11:49:09Z"
    lastUpdateTime: "2025-05-26T11:49:09Z"
    message: 'error during apply of object "monitoring.coreos.com/v1/Prometheus/garden/seed":
      Internal error occurred: failed calling webhook "prometheus.local.extensions.gardener.cloud":
      failed to call webhook: Post "https://gardener-extension-provider-local.extension-provider-local-rq9gk.svc:443/prometheus?timeout=10s":
      tls: failed to verify certificate: x509: certificate signed by unknown authority
      (possibly because of "crypto/rsa: verification error" while trying to verify
      candidate authority certificate "ca-provider-local-webhook-f0e82f96")'
    reason: ApplyFailed
    status: Progressing
    type: ObservabilityComponentsHealthy
```

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Dedicated kustomization for `Garden` manifest

Previously, we were using `example/operator/20-garden.yaml`, but it's
better to split it so that we are able to support multiple scenarios.

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Move `Extension` generation to `dev-setup/extensions/provider-local`

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* `Extension`s for calico and cilium

Deployed by default in operator setup

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Move garden config to `/dev-setup`

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Move `provider-local` `Controller{Registration,Deployment}` to `/dev-setup`

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Use garden config from `/dev-setup`

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* New `gardenlet` skaffold config in `skaffold-operator-garden.yaml`

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Move `Gardenlet` kustomizations to `/dev-setup`

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Switch ha-multi-zone setup to operator-based setup 😱

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Move virtual-garden kubeconfig to /dev-setup/kubeconfigs

Target picture: All kubeconfigs in central place, not scattered across
several (sub-)directories.

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Drop multi-zone setup in `remote-local-setup.yaml`

The operator-setup (which is already supported by the remote local
setup) is equivalent.

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Upgrade tests: Prevent cross-namespace secret ref for `SecretBinding`

* Dedicated cert-based kubeconfig for virtual garden

Without this, when requesting admin/viewer kubeconfigs for a Shoot in
this operator-managed garden, the issued client certificate will always
contain the user name `system:serviceaccount:kube-system:gardener`.
This is because this ServiceAccount was used to request the admin/viewer
kubeconfig.
Since exactly this ServiceAccount is also used by `gardenlet` when
communicating with Shoots, it has cluster-admin access in all shoot
clusters.
This results in viewer kubeconfigs also having cluster-admin access,
breaking the e2e tests.

See for example: https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/12213/pull-gardener-e2e-kind-ha-multi-zone/1929570125401296896

With a dedicated cert-based kubeconfig, we also improve the developers'
lives. The token-based kubeconfig is rotated each `12h`, which is
annoying for development scenarios. This cert-based kubeconfig is valid
for `1y`.

* Deploy nginx-ingress Gateway/VirtualServices when seed is garden

Previously, this was entirely skipped, leading to a bug which prevented
the seed ingress domain from being accessible (in case the seed is the
garden at the same time).

Now that we switch the e2e/dev setups to this scenario, this bug
surfaced, see for example: https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/12213/pull-gardener-e2e-kind-ha-multi-zone/1929562456758161408

---------

Co-authored-by: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>