Ignore shoot Prometheus ManagedResource before patching
Shoot Prometheus instances are actually managed by the GRM. This means
any patches directly applied to the Prometheus custom resources will be
instantly reverted unless the corresponding ManagedResource is ignored
via annotation:

    resources.gardener.cloud/ignore="true"

Upon completion, the ManagedResource can be unignored for the GRM to
reset the Prometheus state. This is done in a deferred call to make sure
the annotation is always removed, also in case of failures.

If the gardenlet process is killed before the cleanup tasks are
complete, the ignore annotation could be leaked. To guarantee the state
is eventually consistent, the gardenlet makes sure to remove ignore
annotations before start up.