[GEP-28] Add support for `NodeAgentAuthorizer` feature to `gardenadm init` (#12169)
* Make `machineNamespace`s optional

* Adapt node-agent-authorizer webhook that it supports a scenario without machines

* Adapt csr-approver that it supports a scenario without machines

* Support mutating static pods during translation

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Support `NodeAgentAuthorizer` feature gate in `gardenadm`

* Drop manual approval of kubelet server certificate in gardenadm flow

This is done by csr-approver now.

* Wait until `gardener-node-agent` is ready

This step verifies, that gardener-node-agent is really running and not in a kind of crash loop.
Additionally, this ensures that kube-apiserver has been rolled and the node-agent-authorizer webhook is accessible. Otherwise, gardener-node-agent would not be able to start.

* Support `gardenadm join`

Machine name must be set correctly to make this work. Additionally, the last-applied-osc.yaml file must not be written when GNA Reconciler applies the init OSC. Otherwise, it would delete the `machine-name` file right away because it is in the init OSC but not in the original one.
While gardener-node-agent is bootstrapping there is no node in the `gardenadm join` scenario. Thus, it requires access to all OSC secrets because node-agent-authorizer does not know which worker pool the new GNA is using. If there would be one OSC per node, we could restrict the access to the OSC secrets which are not used on any node.

* Address PR review feedback

* Enable node-agent-authorizer webhook in two steps

* Address PR review feedback v2

* Activate `gardener-node-agent` early on in the `init` flow

See https://github.com/gardener/gardener/pull/12169#discussion_r2102317143 for more details

- requires temporary cluster-admin permissions since node-agent
  authorizer is not running yet
- GNA now runs before `kubelet` is bootstrapped
- this allows to remove the manual kubelet bootstrapping (GNA does it
  already)
- at the end of the flow, the temporary cluster-admin permissions are
  removed, and it is ensured that GNA is still active/ready

In addition, deploying the control plane deployments now also updates
the OSC and the Secret reconciled by node-agent.

* Address PR review feedback

* Adapt unit tests

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>