[release-v1.121] Enhance Extensions Webhook Library (#12289)
* Drop custom `http.Handler` for remote addr injection

The controller-runtime library supports this out-of-the-box.

* Drop `MutatorWithShootClient`

Generally, the `Mutator` interface must be implemented now.
If a shoot client is needed, then the new `WantsShootClient` interface
can also be implemented. If so, the shoot client is injected into the
context and can be fetched from there.

* Support injection of `Cluster` object

Some extensions might need to read information about the `Shoot`, so
injecting the corresponding `Cluster` object should be supported.

* Remove `handlerAction` interface

Remove `handlerAction` interface to:
- Fix `WantsShootClient` and `WantsClusterObject` interface check
- Simplify code

Fix context usage

* Move interfaces

* Remove `gvkToObject`

Further reduce code complexity by removing the `gvkToObject` map.

* Adapt tests

* Address nits

* Address review feedback

* Address review feedback II

---------

Co-authored-by: rfranzke <rafael.franzke@sap.com>
Co-authored-by: Tim Usner <tim.usner@sap.com>