Split singular Istio CRD manifest into singular manifests
This was an issue with the `CRDDeployer`, as we create a mapping from the name of the CRD to its manifest.
With a singular manifest, we can successfully create the CRDs, but only the first one in the manifest will be deleted, because we delete by CRD name and not the manifest.