Do not add the init container for shoot Prometheus instances
Shoot Prometheus instances have the additional challenge that they can't
be guaranteed to have been successfully reconciled. For instance, if
shoots are hibernated or if the reconciliation fails. In consequence,
previous commit does not suffice.

Previous commit becomes now specific for Prometheus instances in the
garden and seed clusters, where reconciliation must work. A more
sophisticated migration for shoot Prometheus instances is provided in
the following commits.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>