Do not add the init container for shoot Prometheus instances
Shoot Prometheus instances have the additional challenge of not being
guaranteed successful reconciliation. For instance, if shoots are
hibernated or if the reconciliation fails.

In consequence, the previous commit is now tailored specifically for
Prometheus instances in the garden and seed clusters, where
reconciliation is expected to work. A more advanced migration for
shoot Prometheus instances will be addressed in the subsequent commits.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>