Execute shoot Prometheus clean up asynchronously
If there are many hibernated shoots that need to be woken up, new nodes
will need to be spun up, delaying the startup of the gardenlet in few
minutes.

This commit improves this by letting this cleanup run in the background.
Note this is implemented in a way that it can never return an error.
Failures can still happen and errors are properly logged but they are
not propagated back to the gardenlet.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>