Increase timeout for `gardenadm init` to `10m`
After https://github.com/gardener/gardener/pull/12169, there is another
restart of the control plane to activate the NodeAgentAuthorizer webhook
(this changes configuration in `kube-apiserver`). Additionally, we have
to ensure that `gardener-node-agent` is still running afterwards, and
this can take some additional time.

In some e2e runs, the current `5m` timeout is simply too short, see for
example:

- https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/12344/pull-gardener-e2e-kind-gardenadm/1934980463915438080
- https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/12335/pull-gardener-e2e-kind-gardenadm/1934970509573754880
- https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/12318/pull-gardener-e2e-kind-gardenadm/1934964955145048064

or generally
https://prow.gardener.cloud/?repo=gardener%2Fgardener&job=pull-gardener-e2e-kind-gardenadm&state=failure,
if you are fast enough