Do not reset cleanup annotation during reconciliation
We need to preserve when a shoot Prometheus resource is marked as
cleaned up so that gardenlet restarts do not keep attempting to clean up
Prometheus instances that were already done.

When a new shoot Prometheus instance is deployed, the GRM detects the
Prometheus resource does not exist yet and creates it directly with the
cleaned up mark. However, the first time the shoot is reconciled the
Prometheus resource exists and the mark is reverted. This is not the
case for migrated shoot Prometheus instances, where the annotation is
owned by the gardenlet.

To fix this, this commit makes sure the GRM preserves the annotation
across shoot reconciliations.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>