Review: Return errors
Ignoring the errors will guarantee cleaning up the obsolete
"prometheus-" folders won't interfere with the startup of the gardenlet.
However, it won't guarantee that all Prometheus volumes will be
successfully cleaned up once the migration code is removed.

With this commit, we favor ensuring the obsolete folders are correctly
cleaned up and accept the additional overhead in the gardenlet startup.
If an error occurs, now the gardenlet will fail and restart.