Clean up obsolete `prometheus-` folder from Prometheus volumes (#12219)
* Clean up obsolete "prometheus-" folder from Prometheus volumes

In GEP-19, the Prometheus operator was introduced in Gardener to manage
Prometheus instances. This operator configures Prometheus statefulsets
to use the "prometheus-db" folder as a subpath for storage:

    > k get pods prometheus-shoot-0 -o json | jq -r '
        (.spec.volumes[] | select(.name == "prometheus-db")),
        (.spec.containers[].volumeMounts[] | select(.name == "prometheus-db"))'
    {
      "name": "prometheus-db",
      "persistentVolumeClaim": {
        "claimName": "prometheus-db-prometheus-shoot-0"
      }
    }
    {
      "mountPath": "/prometheus",
      "name": "prometheus-db",
      "subPath": "prometheus-db"
    }

Previously, the "prometheus-" folder was used instead, as in:

    https://github.com/gardener/gardener/blob/96335a6487944f05322ac8f182f46c1e738ee2ff/pkg/component/monitoring/charts/seed-monitoring/charts/core/charts/prometheus/templates/prometheus.yaml#L134-L135

During the migration to the Prometheus operator, there was a disk
migration to move data between the two folders:

    https://github.com/gardener/gardener/pull/9128/commits/ca0a1863fca24146d6689bdb6b8d9017a44441d9#diff-c6e1ff35f50162ec0f1[…]1c0d5253120d9c71R101-R106

However, the old "prometheus-" folder still exists in some volumes,
consuming up to e.g., 7GB of data in certain control-plane Prometheus
instances. This poses a risk of Prometheus running out of disk space, as
it operates on a 20GB volume with a disk retention setting of 15GB. In
some cases, we have seen the old "prometheus-" folder as a sibling of
"prometheus-db". In some other cases, as a child.

We fix this leftover by introducing a temporal init container in
Prometheus to remove the obsolete folders.

Unfortunately, identifying the exact sequence of steps
or potential disk migration failures that led to this situation is
challenging after this time and will remain unresolved.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Do not add the init container for shoot Prometheus instances

Shoot Prometheus instances have the additional challenge of not being
guaranteed successful reconciliation. For instance, if shoots are
hibernated or if the reconciliation fails.

In consequence, the previous commit is now tailored specifically for
Prometheus instances in the garden and seed clusters, where
reconciliation is expected to work. A more advanced migration for
shoot Prometheus instances will be addressed in the subsequent commits.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Prepare migration to remove shoot Prometheus obsolete folders

This commit is only the skeleton to clean up shoot Prometheus folders as
part of the startup of the gardenlet. It iterates over the Cluster
resources in the seed to find shoots, but executes nothing yet.

* Ignore shoot Prometheus ManagedResource before patching

Shoot Prometheus instances are actually managed by the GRM. This means
any patches directly applied to the Prometheus custom resources will be
instantly reverted unless the corresponding ManagedResource is ignored
via annotation:

    resources.gardener.cloud/ignore="true"

Upon completion, the ManagedResource can be unignored for the GRM to
reset the Prometheus state. This is done in a deferred call to make sure
the annotation is always removed, also in case of failures.

If the gardenlet process is killed before the cleanup tasks are
complete, the ignore annotation could be leaked. To guarantee the state
is eventually consistent, the gardenlet makes sure to remove ignore
annotations before start up.

* Wait for shoot Prometheus pod to run

This commit prepares the shoot Prometheus cleanup by creating an empty
patch for Prometheus resources, applying it and waiting for shoot
Prometheus pods to be in running state.

The following commits will populate such empty patch.

* Allow the gardenlet to read Prometheus resources

This is necessary for the gardenlet to be able to read and patch
Prometheus resources.

* Patch shoot Prometheus to inject init container

This commit makes sure all shoot Prometheus instances receive the init
container that removes the obsolete "prometheus-" folders. Such init
container is only injected if it is still not present in the Prometheus
resource.

* Patch shoot Prometheus number of replicas to 1

This guarantees the init container runs for hibernated shoots. For
running shoots, setting the number of replicas to 1 is a no-op.

* Execute shoot Prometheus clean up asynchronously

If there are many hibernated shoots that need to be woken up, new nodes
will need to be spun up, delaying the startup of the gardenlet in few
minutes.

This commit improves this by letting this cleanup run in the background.
Note this is implemented in a way that it can never return an error.
Failures can still happen and errors are properly logged but they are
not propagated back to the gardenlet.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Add annotation to avoid reiterating attempts to clean up

This commit uses a temporal annotation to mark shoot Prometheus
instances already cleaned up. This way, the gardenlet doesn't repeatedly
attempt to clean up every shoot Prometheus on every restart.

This temporal annotation will be removed in a second migration step.

* Mark new shoot Prometheus instances as cleaned up

This way, the gardenlet startup does not execute the cleanup for new
shoot Prometheus instances.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Do not reset cleanup annotation during reconciliation

We need to preserve when a shoot Prometheus resource is marked as
cleaned up so that gardenlet restarts do not keep attempting to clean up
Prometheus instances that were already done.

When a new shoot Prometheus instance is deployed, the GRM detects the
Prometheus resource does not exist yet and creates it directly with the
cleaned up mark. However, the first time the shoot is reconciled the
Prometheus resource exists and the mark is reverted. This is not the
case for migrated shoot Prometheus instances, where the annotation is
owned by the gardenlet.

To fix this, this commit makes sure the GRM preserves the annotation
across shoot reconciliations.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Review: Return errors

Ignoring the errors will guarantee cleaning up the obsolete
"prometheus-" folders won't interfere with the startup of the gardenlet.
However, it won't guarantee that all Prometheus volumes will be
successfully cleaned up once the migration code is removed.

With this commit, we favor ensuring the obsolete folders are correctly
cleaned up and accept the additional overhead in the gardenlet startup.
If an error occurs, now the gardenlet will fail and restart.

* Review: Check if the Cluster resource is available in the seed

After returning errors in previous commit, a new error might be returned
if the Cluster resource was not yet installed in the seed when the
gardenlet runs.

* Review: Improve logging while waiting for the pod to be ready

The retry method does not log returned errors between retries. This
commit improves this by logging the error before returning.

* Refactor Error Handling for Unignoring ManagedResource

Ensure proper error handling when unignoring a ManagedResource:

- Log potential unignore errors but return the original error if cleanup
  fails.
- Return the unignore error if cleanup succeeds but unignore fails.

This update removes the defer call to unignore the ManagedResource,
allowing explicit logging or returning of unignore errors as needed. To
enhance code readability and minimize error checks, the unignore
operation is now performed as late as possible.

* Review: Skip shoots being deleted

---------

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>