fix(publicip/test): Include additional `MatchRegexp` to catch inverse errors order
Fixes: https://github.com/gardener/gardener/issues/12469

This commit adds one more `MatchRegexp` matcher that catches that cases
when the `IPv6` error is returned before the `IPv4` one, causing the
current expression:

```
.*IPv4.*: foo.*\n.*IPv6.*: context deadline exceeded.*
```

to fails it's match as it covers only the first scenario.

The additional matcher:

```
.*IPv6.*: context deadline exceeded.*\n.*IPv4.*: foo.*
```

appended with an `Or`, covers the second scenario and resolves
the flaky issue that we've encountered.