chore(validation): Remove redundant `spec` update checks
This commit removes the redundant `Shoot`, `Seed` and `Garden`
validation(s) over the `VerticalPodAutoscaler` configuration when
the resources get updates.

The top-level `Validate(Shoot/Seed/Garden)Update()` functions internally
call `Validate(Shoot/Seed/Garden)()` on the new version of the
resource that perform all the necessary validations.

Even though the `Shoot` and `Seed` spec validation functions

- ValidateShootSpecUpdate
- ValidateSeedSpecUpdate

are exported and can be called regardless of the
`Validate(Shoot/Seed)Update()` function(s), their main purpose is to
prevent changes to `immutable` fields that require both the
`old` and `new` versions of the updating resource.