Validate `spec.addons.nginxIngress.config`
Validating the actual content of the config map values is difficult
because there are a lot of potential valid values and checking them
thoroughly would also cause maintenance effort for each new nginx
controller release as they might change. Therefore, this first just
checks whether the keys are correct and the size is not too large.