Introduce dashboards for Istio L7 load balancing observability (#12509)
* Deploy one service monitor per istio ingress gateway

This allows scraping istio metrics from different prometheus instances (`prometheus-garden` for the virtual garden istio and `prometheus-aggregate` for the seed istios).

* Add `upstream.address` to istio ingress metrics

* Scrape metrics required for the new Istio dashboards

* Consolidate istio dashboards for seed and garden

This works only if both scenarios use the default datasource of plutono everywhere.
Since this is `prometheus-aggregate` in the seeds case, it needs to scrape additional metrics from istio namespaces.

* Introduce istio service and istio workload dashboards