chore(garbagecollector_suite/test): Increase initial garbage collection delay
This commit updates the garbagecollector test suite `BeforeStart` with an
additional MinimumObjectLifetime of 5s as an initial delay for
the Garbage collection of newly created resources.

The reason for the change is that the Garbage collection tests setup
create object relations that might not be configured propertly if
the garbage collection does not take into account an initial delay.

Secret(s)/Configmap(s) that should be referenced within the test setup can end up
being deleted by the garbage collection before their relation to a
Deployment, StatefulSet, etc is created.