Deflake `test/integration/resourcemanager/garbagecollector` integration test (#12578)
* fix(garbagecollector/test): Ensure `referencing resources` are observed by the `resource manager`

This commit instroduces an additional Eventually() block that expects
the Resource Manager to acknowledge the referencing resources before
validating the garbage collection clean up.

Also:
- Increase the Eventually() blocks timeouts to 15 second to leave
  the garbage collection reconciler more time to list all parent
  resources and map the correlation b/w them and the referenced
  secret/configmaps.

* chore(garbagecollector_suite/test): Increase initial garbage collection delay

This commit updates the garbagecollector test suite `BeforeStart` with an
additional MinimumObjectLifetime of 5s as an initial delay for
the Garbage collection of newly created resources.

The reason for the change is that the Garbage collection tests setup
create object relations that might not be configured propertly if
the garbage collection does not take into account an initial delay.

Secret(s)/Configmap(s) that should be referenced within the test setup can end up
being deleted by the garbage collection before their relation to a
Deployment, StatefulSet, etc is created.

* chore(garbagecollector/suite/test): Increase garbage collection `SyncPeriod`

This commit updates the garbage collection `SyncPeriod` from 100ms to 1s

Also:
- Remove the `Eventually()` timeouts of 15s

* chore(garbagecollector/test): Remove usage of `refObj` in loop

This commit removes the usage an in intermediate `refObj` within
the loop that checks for `referencingObjects` existence.

See https://go.dev/blog/loopvar-preview