Clean up annotations of kube-apiserver service to purge old load balancer annotations.
Since quite a while SNI routing is the only supported approach for
kube-apiserver exposure. This means that the kube-apiserver service in
the shoot control plane is a service of type ClusterIP. Unfortunately,
for quite a lot of releases
`.spec.settings.loadbalancerServices.annotations` were still added to
the annotations of the `kube-apiserver` service. This could lead to load
balancer controllers spamming kubernetes events and unnecessarily trying
to reconile a service of type ClusterIP.
With this change, the old annoations will be purged and replaced only
with the needed ones.