Fix cluster overview `cluster-autoscaler` bug
The cluster overview dashboard shows panels with the status of various
cluster components like `etcd`, `kube-apiserver`, `kubelet`s, and more.
Among those is also the `cluster-autoscaler`. However, there are some cases
in which Gardener does not deploy the `cluster-autoscaler`: When the
cluster is a workerless cluster, a version of the dashboard is deployed
without the `kubelet` and `cluster-autoscaler`. However, Gardener also
doesn't deploy the cluster-autoscaler when the minimum and maximum number
of nodes are identical.
In that case, Prometheus fails to scrape the cluster-autoscaler scrape
target, which is mapped to the `DOWN` value in this Plutono dashboard.

With this commit, we also check whether the `cluster-autoscaler` deployment
is supposed to have more than zero replicas, i.e. whether CA pods are
present at all. Now, when CA is *not* deployed and the scrape job fails,
the panel reads `not deployed`. And when CA is deployed, the previous
behavior still applies.