Adapt generic actuator to `gardenadm bootstrap` scenario
Skip all shoot-related operations as documented in the extension contract.

Also, pass `ControlPlane` and `Cluster` to secret funcs.
This allows the provider extension to distinguish which secrets should be deployed based on the scenario.
E.g., this allows detecting the `gardenadm bootstrap` scenario as shown in the local `ControlPlane` controller.
Technically, this is not needed in `shootAccessSecrets` but we add the same params for consistency with `secretConfigs`.