Improve Input Validation of Networking-related `Shoot` fields (#12539)
* Validate `spec.networking.type`

The value is used as part of a label name. Hence, it can only contain
certain characters and has a limited length.

* Validate `spec.systemComponents.coreDNS.rewriting.commonSuffixes`

The common suffixes used for DNS rewriting should resemble domain names
and follow the corresponding rules.

* Validate `spec.addons.nginxIngress.loadBalancerSourceRanges`

The load balancer source ranges should be valid CIDRs.

* Validate `spec.addons.nginxIngress.config`

Validating the actual content of the config map values is difficult
because there are a lot of potential valid values and checking them
thoroughly would also cause maintenance effort for each new nginx
controller release as they might change. Therefore, this first just
checks whether the keys are correct and the size is not too large.

* Validate `spec.exposureClassName` and `ObjectMeta` of ExposureClass
resource.

The validation of `ExposureClass` was missing. It is now implemented and
forces stricter restrictions on the name of an `ExposureClass`.