Configure `vpa-updater` with a `ServiceMonitor` to allow metrics scraping (#12677)
* feat(vpa/updater): add `ServiceMonitor` to `vpa-updater`

This commit introduces a ServiceMonitor resource to the VPA updater
in order to scrape the component metrics by Prometheus.

Also:
- As a requirement, a new Service resource is created to enable the
  ServiceMonitor referrence with the VPA updater.

* feat(vpa/updater/test): include `ServiceMonitor` and `Service` to the expected resources

This commit adds the new vpa-updater ServiceMonitor and Service to the
expected resources within the vpa unit tests.

* feat(vpa/updater/test): add tests to check if `ServiceMonitor` is labeled correctly

* chore(vpa/updater/test): use `switch` when creating objects for different cluster types

This commit instruments the utility functions that create testing
resources to use switch and separate the different cluster types
in dedicated cases.