Introduce new managed-by label for Managed Resources
This label will help identify the component that manages a certain
Managed Resource. For example, in case of Prometheus, the garden and
longterm Prometheus Managed Resources will be labeled with

    managed-by: gardener-operator

while the aggregate, cache, seed and shoot Prometheus with

    managed-by: gardenlet

This is used in the following commits by the gardener-operator and
gardenlet to find the correct list of Prometheus they manage in order to
only query those for health issues. This way, network access can be
minimised to prevent, for example, the gardenlet from querying the
garden Prometheus when the seed is also the runtime cluster.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>