Add a new CoreDNS for translating between IPv6 and IPv4 worlds.
In older releases of docker, domain name resolution was possible via
both IPv4 and IPv6. Unfortunately, with v27 or v28 this is no longer the
case. Docker only provides name resolution via IPv4. For pure IPv6-only
container workloads this is a problem.
This change works around the problem by adding another CoreDNS located
in the host network, which is able to translate between both worlds as
it has both IPv6 and IPv4. The kind CoreDNS is then configured to use
this translation service as upstream DNS provider.