`gardenadm init` stores config dir location
It stores the location of the config directory in a well-known file.
Later, in `gardenadm connect`, we can just read this file and get to
know all the manifests. This improves UX because the user does not need
to care about it when running `gardenadm connect`.

Note: This is just a preparation, the actual functionality (`gardenadm
connect` creating the resources from the config directory in the
Gardener API) will come with a follow-up PR.