Fix Test Machinery Shoot Application Networking Test for IPv6. (#12721)
The test uses nginx and tries to curl between the pods. Unfortunately,
the default configuration of nginx only uses IPv4, i.e. it does not
listen for IPv6 traffic. This causes problems in conjunction with the
fact that the test only uses the first pod IP, which is from the primary
IP family of the cluster. In case the cluster is IPv6-only or has IPv6
as primary address family the call was failing because nobody was
listening on the other end.
In addition to that, the concatenation of IP address and port was only
working for IPv4 addresses. With IPv6, the address needs to be enclosed
in square brackets.
The first issue was addressed by adding an nginx configuration to listen
on IPv6 as well. This also works in IPv4 clusters as IPv6 is always
available locally. The second issue was addressed by using the proper
utility function from the go standard library.