`gardenlet` bootstrap procedure creates CSR for autonomous shoot
Common name: `gardener.cloud:system:shoot:<namespace>:<name>`
Groups: `gardener.cloud:system:shoots`

It reads the `shoot-info` `ConfigMap` in the `kube-system` namespace in
order to get information about the shoot it is running in.

For now, we don't delete the bootstrap auth secret because we don't have
permission for it yet. This will be added in a subsequent commit.