Switch local IPv6 setup to `gardener-operator` (#12709)
* Prepare kind cluster for ipv6 with `gardener-operator`

* Support `gardener-operator` deployment with IPv6

* Support garden cluster with IPv6

* Prepare seed/gardenlet deployment with ipv6

* Switch setup and cleanup no longer needed stuff

* Add a new CoreDNS for translating between IPv6 and IPv4 worlds.

In older releases of docker, domain name resolution was possible via
both IPv4 and IPv6. Unfortunately, with v27 or v28 this is no longer the
case. Docker only provides name resolution via IPv4. For pure IPv6-only
container workloads this is a problem.
This change works around the problem by adding another CoreDNS located
in the host network, which is able to translate between both worlds as
it has both IPv6 and IPv4. The kind CoreDNS is then configured to use
this translation service as upstream DNS provider.

* Allow test to run both in operator and non-operator setup

* Fix seed garden access test

https://github.com/gardener/gardener/blob/485c25124ea536e4610c627109017dd34d434921/test/e2e/gardener/seed/renew_garden_access_secrets.go#L82 uses the wrong client when the operator setup is used

Ref https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/12709/pull-gardener-e2e-kind-ipv6/1956070713743904768

* Fix Gardener Seed tests by using the proper cluster client.

Garden client and seed client are different in the operator-based setup.

* Initialize Seed Context for End-to-End Tests

* Further cleanup of no longer needed files/config

* Service controller should not mix responsibility of istio deployments

Without this, `gardener-operator` also reconciles the istios deployed by
`gardenlet`. While this was no issue before, now it is with the IPv6
setup. `gardener-operator` conflicts with what `provider-local` tries to
patch into the `istio-ingress/istio-ingressgateway` `Service`.

Similarly, provider-local should not tamper with the
`virtual-garden-istio-ingress/istio-ingressgateway`.

Hence, let's better restrict them to what they are really responsible for.

---------

Co-authored-by: Johannes Scheerer <johannes.scheerer@sap.com>