Add Prometheus health checking to HealthChecker
This adds three new methods to HealthChecker:

- CheckPrometheus() and checkPrometheusReplicaHealth() validate that
  there are no health issues for a given Prometheus resource.
  CheckPrometheus() extracts the number of Prometheus replicas (i.e.,
  pods) and calls checkPrometheusReplicaHealth() on each pod.
  checkPrometheusReplicaHealth() invokes the actual call to query a
  Prometheus endpoint.
- CheckPrometheuses() uses the first methods to check for Prometheus
  health issues from a list of Prometheus. This method provides a common
  entry point for both gardener-operator and gardenlet to validate
  Prometheus health.

Prometheus resources and their respective replicas are checked in
parallel to prevent timeouts from delaying sequential health checks.

Furthermore, Prometheus resources are sorted by namespace and name
before health checking. This guarantees the same failing condition is
returned when there are multiple to report. Without this, regular
condition status checks could return different failing conditions
depending on the order, and cause frequent writes in etcd.

A new option field is added to HealthChecker to provide a
PrometheusHealthChecker abstraction that enables tests to implement
their own PrometheusHealthChecker function. This approach allows for
testing the multi-replica orchestration without the need for a test HTTP
server.

A test to validate Prometheus resources and replicas are checked in
parallel is included.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>