Temporarily disable monitoring for Istio Ingress Gateways
Mitigates the metrics leak issue until 12699 is resolved.

By skipping the creation of the ServiceMonitor and Telemetry resources,
monitoring for the Istio Ingress Gateways is effectively disabled.

Note that the managed resource is still created so that the
gardener resource manager can handle the deletion of these artifacts.

The code change is kept minimal to allow for easy revert
once the issue is fixed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>
Co-authored-by: Christoph Kleineweber <c.kleineweber@sap.com>