[GEP-28] `gardenadm bootstrap`: enable provider-local DNS config (#12836)
* Enable provider-local-coredns in `gardenadm bootstrap`

Don't enable it within the autonomous shoot cluster itself (both scenarios)
Also, don't mutate high-touch machine pods when deploying both scenarios on the same cluster simultaneously (e.g., in CI).

* Replace explicit `NetworkPolicies` with generated ones

Instead of explicitly creating `NetworkPolicies` in the `garden` and shoot namespaces, we enable the `NetworkPolicy` controllers
on the `gardener-extension-provider-local-coredns` namespace, and generate `NetworkPolicies` in the `garden` and shoot namespaces.
The `dnsconfig` webhook injects the corresponding labels along with the DNS configuration itself.
For enabling the `NetworkPolicy` controllers, we also need to explicitly allow the provider-local coredns to talk to the cluster's
coredns in `kube-system`.

We're doing this now, because the explicit `NetworkPolicy` `allow-to-provider-local-coredns` cannot be deployed to `garden` namespace
because it doesn't exist in `gardenadm bootstrap`.

* Add workaround `NetworkPolicy` for broken multi-node tests