Add missing `configmaps` resource to `seed-restriction` webhook config (#12858)
* Add missing `configmaps` resource to `seed-restriction` webhook config

* Allow `gardener-admission-controller` to list/watch `ConfigMap`s

* Do not enable `configmaps` resource in deprecated `gardener/controlplane` chart

When using the `gardener/controlplane` chart, all components
(`gardener-apiserver`, `gardener-admission-controller`, etc.) are
deployed in parallel. When the seed-restriction webhook also acts on
`ConfigMap`s, this is a problem. `gardener-apiserver` creates
a `ConfigMap` when it starts before it gets ready, but
`gardener-admission-controller` cannot start when `gardener-apiserver`
is not ready.

As the chart is already deprecated and planned for removal in a couple
of months, I guess it's best to continue living with this deficiency.