Adapt tests
This was a fun one...
The old package did some random newlines, whilst the new ones cannot do this.

Once again, we cannot use the new v2, because of the unstable sorting issue.

There is an _internal_ option in the parser to set a maximum width, but these are unexported options. ¯\_(ツ)_/¯