[GEP-28] Make `SeedConfig` optional in `gardenlet`'s component configuration (#12871)
* No longer default `SeedConfig` if unset

* No longer fail if `SeedConfig` is unset

* No longer require a seed config in validation code

* Enforce `seedTemplate` where required

In `ManagedSeed`s and in `gardener-operator`'s `Gardenlet` controller,
we still expect `seedTemplate`s in the gardenlet component config.

In the future, we'll adapt the `app/options.go` to only require the
`SeedConfig` when deployed into a seed cluster, but not when deployed
into an autonomous shoot cluster.

* Adapt `gardenletdeployer` controller

Note: In a future PR, we will use this controller in the `gardenadm
connect` command in order to deploy the `gardenlet` into the autonomous
shoot cluster.

* Remove no longer needed `inTemplate` argument