[GEP-28] Implement custom verb for marking `Shoot`s as autonomous, and implement validation for `.spec.provider.workers[].controlPlane` (#12885)
* Prefactor: Only pass `[]Worker`s to helper funcs

This makes it easier later in the validation code when we don't have the
full `Shoot` object at hand but only the `.spec.provider` or the list of
workers.

* Adapt usages

* Add similar helper functions for the internal API version

We will use them in a later commit in the `customverbauthorizer`
admission plugin and the validation code.

* Harden validation for autonomous shoots

* Introduce custom `mark-autonomous` verb for `Shoot`s

* Grant `mark-autonomous` verb to Gardener admins

End-users are not granted with this custom verb for now, since GEP-28
only targets the 'Gardener operators' use case and end-users are
out-of-scope.

* Address PR review feedback