feat(docs): Add feature gate and usage guide for enabling in-place Pod resource updates
This commit udpates the feature gates document with an entry for the
gardenlet and gardener operator feature gate, enabling the deployment
of the webhook mutating VPA resources.

In addition, a new in-place updates enablement guide is added to
document the process of enabling the new feature gate and configuring
the `Shoot` and `Seed` cluster's Vertical Pod Autoscaler setups.