Initialize botanist with DNS info
With this commit, the botanist no longer uses a hard-coded internal domain. Instead, the external domain is read from `Shoot.spec.dns.domain`.
There is no internal domain for autonomous shoots, only an external domain.
However, it might be managed outside gardener (high-touch). In this case, the field is only for informing `gardenadm` about the domain for secret generation, etc.